Refactor to `exceptions` module and do not configure logging
